test: assertConvertedPacket を追加
